last tasks for integrity process

+++ device-merchant-service/src/main/java/com/abdali/microhps/devicemerchantservice/repository/MerchantRepository.java
@@ -10,13 +10,14 @@ import org.springframework.data.repository.query.Param;
 import com.abdali.microhps.devicemerchantservice.model.Merchant;
 
 public interface MerchantRepository extends JpaRepository<Merchant, Integer> {
+	
 		// Merchant Not Found
 		@Query("select count(m.id) FROM Merchant m where m.merchantNumber = :MerchantNumber AND m.status NOT IN :Status")
 		Integer merchantStatus(@Param("MerchantNumber") Long merchantNumber, @Param("Status") List<String> status);
 		  
 		Optional<Merchant> findByMerchantNumber(Long merchantNumber);
 		
-		List<Merchant> findAllByDeviceId(Long idDevice);
+		List<Merchant> findAllByDeviceId(Integer idDevice);
 		
 		
 }
