+++ device-merchant-service/src/main/java/com/abdali/microhps/devicemerchantservice/service/MerchantService.java
@@ -14,7 +14,7 @@ public interface MerchantService {
 	
 	MerchantDto findByMerchantNumber(Long merhantNumber);
 	
-	List<MerchantDto> findAllMerchantByIdDevice(Long idDevice);
+	List<MerchantDto> findAllMerchantByIdDevice(Integer idDevice);
 	
 	Boolean merchantCheckStatus(Long merchantNumber, List<String> status);
 }
